services/sa-ml/requirements.txt
aiofiles==22.1.0
aiohttp==3.8.3
aiokafka==0.8.0
aiosignal==1.3.1
anyio==3.6.2
async-timeout==4.0.2
attrs==22.2.0
Brotli==1.0.9
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
fastapi==0.88.0
frozenlist==1.3.3
gevent==22.10.2
geventhttpclient==2.0.2
greenlet==2.0.1
grpcio==1.51.1
h11==0.14.0
idna==3.4
joblib==1.2.0
kafka-python==2.0.2
mlserver==1.2.1
mlserver-sklearn==1.2.1
multidict==6.0.4
numpy==1.24.1
orjson==3.8.3
packaging==22.0
pandas==1.5.2
prometheus-client==0.15.0
protobuf==4.21.12
py-grpc-prometheus==0.7.0
pydantic==1.10.4
python-dateutil==2.8.2
python-dotenv==0.21.0
python-rapidjson==1.9
pytz==2022.7
scikit-learn==1.2.0
scipy==1.9.3
six==1.16.0
sniffio==1.3.0
starlette==0.22.0
starlette-exporter==0.14.0
threadpoolctl==3.1.0
tritonclient==2.29.0
typing_extensions==4.4.0
uvicorn==0.20.0
uvloop==0.17.0
yarl==1.8.2
zope.event==4.6
zope.interface==5.5.2
